% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raster_layer_value.R
\name{add_linz_key}
\alias{add_linz_key}
\title{Add LINZ API key to environment variables}
\usage{
add_linz_key(key)
}
\arguments{
\item{key}{character; LINZ API key. An account is needed to obtain an API
key. The function \code{create_linz_key()} will open a browser window to
create an account and obtain an API key.}
}
\description{
Add LINZ API key to environment variables
}
