% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aeme_parameters_bgc}
\alias{aeme_parameters_bgc}
\title{Example dataframe used for calibrating the biogeochemistry in the AEME
models.}
\format{
\subsection{\code{aeme_parameters_bgc}}{

A data frame with 30 rows and 7 columns:
\describe{
\item{model}{Model for the parameter}
\item{file}{File in which the parameter is stored}
\item{name}{Name of the parameter}
\item{value}{Value of the parameter}
\item{min}{Minimum range of the parameter}
\item{max}{Maximum range of the parameter}
\item{var}{Maximum range of the parameter}
}
}
}
\source{
Package development.
}
\usage{
aeme_parameters_bgc
}
\description{
An example dataframe used for inputting and calibrating AEME models.
}
\keyword{datasets}
