% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemetools-package.R
\docType{package}
\name{aemetools-package}
\alias{aemetools}
\alias{aemetools-package}
\title{aemetools: Tools to expand the Aquatic Ecosystem Model Ensemble Package}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Calibrate the AEME model ensemble, download ERA5-Land data for any point in New Zealand, and run simple hydrological models to generate inflow inputs for the AEME model ensemble.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://limnotrack.com/aemetools/}
  \item Report bugs at \url{https://github.com/limnotrack/aemetools/issues}
}

}
\author{
\strong{Maintainer}: Tadhg Moore \email{tadhg@limnotrack.com} (\href{https://orcid.org/0000-0002-3834-8868}{ORCID})

Authors:
\itemize{
  \item Tadhg Moore \email{tadhg@limnotrack.com} (\href{https://orcid.org/0000-0002-3834-8868}{ORCID})
  \item Chris McBride \email{chris@limnotrack.com}
}

}
\keyword{internal}
