% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_request.R
\name{api_request}
\alias{api_request}
\title{Request data from the LERNZMP API}
\usage{
api_request(
  api_url,
  endpoint,
  query = list(),
  api_key = NULL,
  headers = list()
)
}
\arguments{
\item{api_url}{character; base URL of the API.}

\item{endpoint}{character; specific API endpoint to call.}

\item{query}{list; named list of query parameters.}

\item{api_key}{character; API key for authentication. If NULL, will look for
the key in the LERNZMP_KEY environment variable. If that is not set, will
throw an error.}

\item{headers}{list; additional headers to include in the request. Defaults
to an empty list.}
}
\value{
The raw response object from httr2; parsing is left to the caller.
}
\description{
Handles authentication, error checking, and returns the raw response.
}
