% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_GR.R
\name{calib_GR}
\alias{calib_GR}
\title{Calibrate GR model}
\usage{
calib_GR(
  hydro_model,
  warmup = NULL,
  run_index = NULL,
  FUN_CRIT = airGR::ErrorCrit_NSE,
  FUN_CALIB = airGR::Calibration_Michel,
  IniStates = NULL,
  IniResLevels = NULL
)
}
\arguments{
\item{hydro_model}{\code{\linkS4class{HydroModel}}; object created by \code{\link{make_GR_inputs}()}.}

\item{warmup}{integer vector; row indices of \code{hydro_model@data} to use as
the warm-up period. Defaults to \code{1:(hydro_model@start - 1)}.}

\item{run_index}{integer vector; row indices of \code{hydro_model@data} to use
during calibration. Defaults to
\code{hydro_model@start:nrow(hydro_model@data)}.}

\item{FUN_CRIT}{[function (atomic or list)] error criterion function (e.g. \code{\link[airGR]{ErrorCrit_RMSE}}, \code{\link[airGR]{ErrorCrit_NSE}})}

\item{FUN_CALIB}{(optional) [function] calibration algorithm function (e.g. Calibration_Michel), default = \code{Calibration_Michel}}

\item{IniStates}{(optional) [numeric] object of class \code{IniStates} [mm and °C], see \code{\link[airGR]{CreateIniStates}} for details}

\item{IniResLevels}{(optional) [numeric] vector of initial fillings for the GR stores (4 values; use NA when not relevant for a given model) [- and/or mm]. See details}
}
\value{
An \code{airGR} \code{OutputsCalib} list.
}
\description{
Calibrate a GR hydrological model from the \code{airGR} package using a
\code{\linkS4class{HydroModel}} object.
}
