% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raster_layer_value.R
\name{create_linz_key}
\alias{create_linz_key}
\title{Create LINZ API key}
\usage{
create_linz_key()
}
\description{
This function will open a browser window to create a LINZ account and obtain
an API key. The API key will be copied to the clipboard and can be added to
the environment variables using the \code{add_linz_key()} function.
}
