% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-calib_helpers.R
\name{decode_param}
\alias{decode_param}
\title{Decode a full parameter name to extract the base name.}
\usage{
decode_param(name_full)
}
\arguments{
\item{name_full}{The full parameter name in the format "group/name[index]".}
}
\value{
The base parameter name without group and index.
}
\description{
Decode a full parameter name to extract the base name.
}
