% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-calib_helpers.R
\name{decode_param_full}
\alias{decode_param_full}
\title{Decode a full parameter name into group, name, and index.}
\usage{
decode_param_full(name_full)
}
\arguments{
\item{name_full}{The full parameter name in the format
"group/.../name[index]".}
}
\value{
A data.frame with columns: group, name, index.
}
\description{
Decode a full parameter name into group, name, and index.
}
