% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-calib_helpers.R
\name{display_param_name}
\alias{display_param_name}
\title{Display parameter information in a human-readable format.}
\usage{
display_param_name(group, name, index)
}
\arguments{
\item{group}{The group name (can be NA).}

\item{name}{The parameter name.}

\item{index}{The index of the parameter (can be NA).}
}
\description{
Display parameter information in a human-readable format.
}
