% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_era5_grib.R
\name{download_era5_grib}
\alias{download_era5_grib}
\title{Download ERA5 GRIB files}
\usage{
download_era5_grib(
  shape = NULL,
  lat,
  lon,
  buffer = 0.1,
  variable = c("10m_u_component_of_wind", "10m_v_component_of_wind",
    "2m_dewpoint_temperature", "2m_temperature", "snowfall", "surface_pressure",
    "surface_solar_radiation_downwards", "surface_thermal_radiation_downwards",
    "total_precipitation"),
  year = 2022,
  month = 1:2,
  site = "test",
  user = NULL,
  era5_dataset = "reanalysis-era5-land",
  path = "."
)
}
\arguments{
\item{shape}{sf object; shapefile or spatial object to download data for.}

\item{lat}{numeric; latitude}

\item{lon}{numeric; longitude}

\item{buffer}{numeric; buffer around the point in degrees. Default is 0.1.}

\item{variable}{vector of ERA5 variable names e.g.
"2m_temperature", "total_precipitation". Default variables are:
"10m_u_component_of_wind", "10m_v_component_of_wind",
"2m_dewpoint_temperature", "2m_temperature", "snowfall",
"surface_pressure", "surface_solar_radiation_downwards",
"surface_thermal_radiation_downwards", "total_precipitation".}

\item{year}{numeric; year or vector of years.}

\item{month}{numeric; month or vector of months. Defaults to 1:12.}

\item{site}{string of site name which will be appended to the file}

\item{user}{user (email address) used to sign up for the ECMWF data service,
if only a single user is needed it defaults to ("ecmwfr").}

\item{era5_dataset}{string; of which ERA5 dataset to use. Can be 'reanalysis-era5-single-levels' or 'reanalysis-era5-land'}

\item{path}{filepath to store downloaded file}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{download_era5_grib()} is superseded by
\code{metscale::download_era5_cds()}. ERA5 extraction, conversion and
bias-correction have moved to the
\href{https://github.com/limnotrack/metscale}{metscale} package; install it
with \code{remotes::install_github("limnotrack/metscale")}.

Download ERA5 meteorological data from the Copernicus Data Store (CDS).
Create a free CDS user account by \href{https://cds.climate.copernicus.eu/user/register}{self
registering}.
Once your user account has been verified you can get your personal user ID
and key by visiting the \href{https://cds.climate.copernicus.eu/user}{user
profile}. Further information can be found on the
\href{https://bluegreen-labs.github.io/ecmwfr/#use}{ecmwfr} package page.
}
\examples{
\dontrun{
lat <- -38.07782
lon <- 176.2673
year <- 2024
month <- 1:2
variable <- "2m_temperature"
path <- "data/test"
ecmwfr::wf_set_key(key = "XXXXXXXX-XXXX-XXXX-XXXX-XXXXXXXXXXXX",
                   user = "person@email.com")
files <- download_era5_grib(lat = lat, lon = lon, year = year, month = month,
                            variable = variable, path = path, user = user)
files
}
}
