% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_parameters_shiny.R
\name{edit_parameters_shiny}
\alias{edit_parameters_shiny}
\title{Edit model parameters using a Shiny gadget}
\usage{
edit_parameters_shiny(param)
}
\arguments{
\item{param}{dataframe; of parameters read in from a csv file. Requires the
columns c("model", "file", "name", "value", "min", "max", "log")}
}
\value{
A data frame with the edited parameters. The structure of the data
frame is the same as the input, with columns for model, file, name, value,
min, max, group and index. If the user cancels the editing, the original
input is returned.
}
\description{
This function launches a Shiny gadget that allows users to edit model
parameters in an interactive table. The parameters are filtered by model and
file, and users can edit the values, minimums, maximums, and groups of the
parameters. The edited parameters are returned as a data frame when the user
clicks "Done". If the user clicks "Cancel", the original parameters are
returned.
}
\examples{
\dontrun{
data("aeme_parameters", package = "AEME")
param <- aeme_parameters
edited_param <- edit_parameters_shiny(param)
}
}
