% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-calib_helpers.R
\name{encode_param}
\alias{encode_param}
\title{Encode a parameter name with its group and index.}
\usage{
encode_param(group, name, index)
}
\arguments{
\item{group}{The group name.}

\item{name}{The parameter name.}

\item{index}{The index of the parameter.}
}
\value{
A string in the format "group/name[index]".
}
\description{
Encode a parameter name with its group and index.
}
