% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{era5_ref_table}
\alias{era5_ref_table}
\title{ERA5 reference table}
\format{
\subsection{\code{era5_ref_table}}{

A data frame with 9 rows and 6 columns:
\describe{
\item{variable}{Variable name}
\item{era5}{ERA5 variable names}
\item{nc}{ERA nc variable names}
\item{ler}{LakeEnsemblR variable names}
\item{aeme}{AEME variable names}
\item{nc}{ERA nc variable names}
}
}
}
\source{
Package development.
}
\usage{
era5_ref_table
}
\description{
A reference table for variable names between ERA5, LER and AEME.
}
\keyword{datasets}
