% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aeme.R
\name{get_aeme}
\alias{get_aeme}
\title{Get Aeme object from LimnoTrack API}
\usage{
get_aeme(id = "LID1", api_url = "https://api.limnotrack.com", api_key = NULL)
}
\arguments{
\item{id}{character or numeric; the character Lernzmp ID (e.g. "LID 1") of
the lake or the numeric FENZ ID (e.g. 1).}

\item{api_url}{character; base URL of the API.}

\item{api_key}{character; API key for authentication. If NULL, will look for
the key in the LERNZMP_KEY environment variable. If that is not set, will
throw an error.}
}
\value{
An Aeme object.
}
\description{
This function retrieves an Aeme object from the LimnoTrack API using the provided
lake ID. The Aeme object is returned as an R object.
}
\examples{
aeme <- get_aeme(id = "LID1", api_key = Sys.getenv("LERNZMP_KEY"))
aeme
}
