% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_params.R
\name{get_all_params}
\alias{get_all_params}
\title{Get all parameter values from calibration results}
\usage{
get_all_params(calib, na_value, fit_col = "fit")
}
\arguments{
\item{calib}{A list with the calibration results loaded using
\code{\link{read_calib}}.}

\item{na_value}{A numeric value which corresponds to the NA value used in
the calibration.}

\item{fit_col}{character; name of column containing fit values. Default is
\code{"fit"}.}
}
\value{
A data frame with all the parameter values.
}
\description{
Get all parameter values from calibration results
}
