% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_best_params.R
\name{get_best_params}
\alias{get_best_params}
\title{Get best parameter values from calibration results}
\usage{
get_best_params(calib, na_value, fit_col = "fit", quantile_threshold = 0.1)
}
\arguments{
\item{calib}{dataframe; output from \code{\link{read_calib}}}

\item{na_value}{A numeric value which corresponds to the NA value used in
the calibration.}

\item{fit_col}{character; name of column containing fit values. Default is
\code{"fit"}.}

\item{quantile_threshold}{Numeric. A value between 0 and 1 specifying the
quantile threshold for selecting the best parameter values based on their fit
values. For example, a value of 0.1 will select parameter values with fit
values in the lowest 10\% of the distribution.}
}
\value{
A data frame with the parameter values.
}
\description{
Get best parameter values from calibration results
}
