% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_catchment_data.R
\name{get_catchment_data}
\alias{get_catchment_data}
\title{Get Catchment Data}
\usage{
get_catchment_data(
  id = 3,
  api_url = "https://api.limnotrack.com",
  api_key = NULL
)
}
\arguments{
\item{id}{character or numeric; the character Lernzmp ID (e.g. "LID 1") of
the lake or the numeric FENZ ID (e.g. 1).}

\item{api_url}{character; base URL of the API.}

\item{api_key}{character; API key for authentication. If NULL, will look for
the key in the LERNZMP_KEY environment variable. If that is not set, will
throw an error.}
}
\value{
A list of sf objects representing the catchment data.
Each element in the list corresponds to a different catchment feature. These
are: catchment boundary ("catchment"), streams ("reaches"), lakes ("lakes),
subcatchments ("subcatchments"), and land cover from the Land Cover DataBase
("lcdb").
}
\description{
Get catchment data from the API and convert GeoJSON elements to sf objects.
}
\examples{
catchment <- get_catchment_data(id = 3, 
api_key = Sys.getenv("LERNZMP_KEY"))
names(catchment)
}
