% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dem_raster.R
\name{get_dem_raster}
\alias{get_dem_raster}
\title{Get DEM raster from LINZ}
\usage{
get_dem_raster(x, lake = NULL, zoom = 14, verbose = FALSE, prompt = FALSE)
}
\arguments{
\item{x}{sf object}

\item{lake}{sf object; of the lake shape. Optional. Default is NULL.}

\item{zoom}{integer; zoom level for the raster. Default is 14.}

\item{verbose}{if TRUE, tiles filepaths, zoom level and
attribution are displayed.}

\item{prompt}{logical; whether to prompt the user to select the DEM layer.
Default is FALSE. This can be useful if the user wants to cycle through the
available DEM layers to find the best one.}
}
\value{
SpatRaster object
}
\description{
Get DEM raster from LINZ
}
\details{
This function will get the DEM raster from LINZ for the given
shape.
}
