% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_depth_contours.R
\name{get_depth_contours}
\alias{get_depth_contours}
\title{Get Depth Contour Data}
\usage{
get_depth_contours(
  id = 1,
  api_url = "https://api.limnotrack.com",
  api_key = NULL
)
}
\arguments{
\item{id}{character or numeric; the character Lernzmp ID (e.g. "LID 1") of
the lake or the numeric FENZ ID (e.g. 1).}

\item{api_url}{character; base URL of the API.}

\item{api_key}{character; API key for authentication. If NULL, will look for
the key in the LERNZMP_KEY environment variable. If that is not set, will
throw an error.}
}
\value{
A sf object representing the depth contours of the lake. The sf object will have
columns for id, name_final, depth, and geometry. The depth column will contain the depth
values for each contour line.
}
\description{
Get depth contour data from the API and convert GeoJSON elements to sf objects.
}
\examples{
depth_contours <- get_depth_contours(id = 1, 
api_key = Sys.getenv("LERNZMP_KEY"))
names(depth_contours)
}
