% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_era5_point.R
\name{get_era5_point}
\alias{get_era5_point}
\title{Get ERA5 data for a coordinate}
\usage{
get_era5_point(
  lat,
  lon,
  years,
  variables = c("MET_tmpair", "MET_tmpdew", "MET_wnduvu", "MET_wnduvv", "MET_pprain",
    "MET_ppsnow", "MET_prsttn", "MET_radswd"),
  format = "aeme",
  parallel = TRUE,
  ncores
)
}
\arguments{
\item{lat}{numeric; Latitude}

\item{lon}{numeric; Longitude}

\item{years}{numeric; vector of years in numeric form to be extracted.
Currently years are limited to 1980-2022.}

\item{variables}{vector; with AEME meteorological variable names to be
downloaded. Defaults to all available variables: c("MET_tmpair",
"MET_tmpdew", "MET_wnduvu", "MET_wnduvv", "MET_pprain", "MET_ppsnow",
"MET_prsttn", "MET_radswd").}

\item{format}{string; the format for column headers. Can be "aeme" or "ler".
Defaults to "aeme".}

\item{parallel}{boolean; parallelise the download of ERA5 variables. Defaults
to TRUE.}

\item{ncores}{integer; number of cores to use for parallelisation. If
missing, defaults to \code{min(c(parallelly::availableCores(omit = 1), length(variables)))}}
}
\value{
dataframe of daily ERA5 data.
}
\description{
Extract \href{https://www.ecmwf.int/en/era5-land}{ERA5-Land} meteorological data
for the closest grid to a particular latitude and longitude for a selected
number of years (1980-2023).
}
