% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lake_shape.R
\name{get_lake_shape}
\alias{get_lake_shape}
\title{Get lake shape as sf object}
\usage{
get_lake_shape(api_url = "https://api.limnotrack.com", id, api_key = NULL)
}
\arguments{
\item{api_url}{character; base URL of the API.}

\item{id}{character or numeric; the character Lernzmp ID (e.g. "LID 1") of
the lake or the numeric FENZ ID (e.g. 1).}

\item{api_key}{character; API key for authentication. If NULL, will look for
the key in the LERNZMP_KEY environment variable. If that is not set, will
throw an error.}
}
\value{
an sf object representing the lake shape.
}
\description{
This function retrieves the shape of a lake from the LimnoTrack API
using the provided Lernzmp ID. The shape is returned as an sf object.
}
\examples{
lake <- get_lake_shape(id = 1, api_key = Sys.getenv("LERNZMP_KEY"))
}
