% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_layer_ids.R
\name{get_layer_ids}
\alias{get_layer_ids}
\title{Get layer IDs for a given spatial object}
\usage{
get_layer_ids(x, type = c("dem", "aerial"))
}
\arguments{
\item{x}{object of class \code{sf}, \code{sfc} or \code{sfg}}

\item{type}{A character vector of length 1 or 2, specifying the type of
layers to return. The default is c("dem", "aerial"). If only one type is
specified, the function will only return layers of that type.}
}
\value{
A data frame with the layer IDs of the layers that intersect with
the spatial object.
}
\description{
This function takes a spatial object and returns the layer IDs for the
aerial imagery and DEM layers that intersect with the spatial object.
}
