% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_linz_sf.R
\name{get_linz_sf}
\alias{get_linz_sf}
\title{Get LINZ data as sf object}
\usage{
get_linz_sf(url, layer_id, key = NULL)
}
\arguments{
\item{url}{character, URL to LINZ API}

\item{layer_id}{numeric; layer ID value for the raster layer on the LINZ
data service. See \url{https://data.linz.govt.nz/}}

\item{key}{character; LINZ API key. This can be set as an environment variable
using the \code{add_linz_key()} function or passed as a character.
See \code{?add_linz_key} for more information with setting up the API key.}
}
\value{
sf object
}
\description{
Get LINZ data as sf object
}
