% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-calib_helpers.R
\name{get_pareto_front}
\alias{get_pareto_front}
\title{Get the Pareto front from a data frame based on specified objective columns.}
\usage{
get_pareto_front(df, obj_cols)
}
\arguments{
\item{df}{A data frame containing the results of a multi-objective optimization.}

\item{obj_cols}{A character vector of column names in df that represent the
objective values to be minimized. The function will identify the rows that are
not dominated by any other row in terms of these objective values.}
}
\value{
A data frame containing only the rows that are on the Pareto front.
}
\description{
Get the Pareto front from a data frame based on specified objective columns.
}
