% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raster_layer_value.R
\name{get_raster_layer_value}
\alias{get_raster_layer_value}
\title{Get raster layer value for a given latitude and longitude}
\usage{
get_raster_layer_value(lat, lon, layer_id, key = NULL)
}
\arguments{
\item{lat}{numeric; Latitude}

\item{lon}{numeric; Longitude}

\item{layer_id}{numeric; layer ID value for the raster layer on the LINZ
data service. See \url{https://data.linz.govt.nz/}}

\item{key}{character; LINZ API key. This can be set as an environment variable
using the \code{add_linz_key()} function or passed as a character.
See \code{?add_linz_key} for more information with setting up the API key.}
}
\value{
numeric value or NA if outside extent
}
\description{
Get raster layer value for a given latitude and longitude
}
