% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raster_tile.R
\name{get_raster_tile}
\alias{get_raster_tile}
\title{Get a raster tile from LINZ}
\usage{
get_raster_tile(x, layer_id, zoom = 15, key = NULL, verbose = FALSE)
}
\arguments{
\item{x}{sf, sfc, bbox, SpatRaster, SpatVector or SpatExtent object.
If \code{x} is a SpatExtent it
must express coordinates in lon/lat WGS84 (epsg:4326).}

\item{layer_id}{numeric; layer ID value for the raster layer on the LINZ
data service. See \url{https://data.linz.govt.nz/}}

\item{zoom}{numeric; zoom level for the tile. Default is 15.}

\item{key}{character; LINZ API key. If NULL, will look for the key in the
LINZ_KEY environment variable. If that is not set, will throw an error.
Use the \code{add_linz_key} function to set the key.}

\item{verbose}{if TRUE, tiles filepaths, zoom level and
attribution are displayed.}
}
\value{
a raster object
}
\description{
Get a raster tile from LINZ
}
