% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sim_params.R
\name{get_sim_params}
\alias{get_sim_params}
\title{Get parameter values from calibration results}
\usage{
get_sim_params(calib, fit_col = "fit", na_value = NULL, quantile_threshold = 1)
}
\arguments{
\item{calib}{A list with the calibration results loaded using
\code{\link{read_calib}}.}

\item{fit_col}{character; name of column containing fit values. Default is
\code{"fit"}.}

\item{na_value}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Numeric. Penalty value
substituted for \code{NA} fit values, this is no longer needed as NA values
are now written to simulation_data in output of calib_aeme() and sa_aeme().
The argument will be removed in a future version.}

\item{quantile_threshold}{Numeric. A value between 0 and 1 specifying the
quantile threshold for selecting the best parameter values based on their fit
values. For example, a value of 0.1 will select parameter values with fit
values in the lowest 10\% of the distribution. Default is 1 (no filtering).}
}
\value{
A data frame with all the parameter values.
}
\description{
Retrieves parameter values from calibration results, with optional filtering
based on fit values. This function is useful for extracting all parameter
sets used in the calibration process, and can be filtered to include only
those with fit values below a specified quantile threshold.
}
