% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_upstream_rec.R
\name{get_upstream_rec}
\alias{get_upstream_rec}
\title{Get all upstream reaches}
\usage{
get_upstream_rec(HydroID, reaches)
}
\arguments{
\item{HydroID}{numeric; identifier for the target reach}

\item{reaches}{sf dataframe; with all reaches with HydroID's}
}
\value{
sf dataframe of all upstream reaches
}
\description{
Get all upstream reaches
}
