% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wlevel.R
\name{get_wlevel}
\alias{get_wlevel}
\title{Get modelled water level}
\usage{
get_wlevel(lake_dir, model, nlev, return_df = FALSE)
}
\arguments{
\item{lake_dir}{lake directory}

\item{model}{string; model name.}

\item{nlev}{numeric; number of levels to extract.}
}
\value{
dataframe of water level.
}
\description{
Get modelled water level
}
