% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gotm_wet_parameters}
\alias{gotm_wet_parameters}
\title{Example dataframe used for calibrating the biogeochemistry in the GOTM-WET
model.}
\format{
\subsection{\code{gotm_wet_parameters}}{

A data frame with 182 rows and 7 columns:
\describe{
\item{model}{Model for the parameter}
\item{file}{File in which the parameter is stored}
\item{name}{Name of the parameter}
\item{value}{Value of the parameter}
\item{min}{Minimum range of the parameter}
\item{max}{Maximum range of the parameter}
\item{var}{Maximum range of the parameter}
}
}
}
\source{
Package development.
}
\usage{
gotm_wet_parameters
}
\description{
All the parameters within the fabm.yaml file. This includes three
phytoplankton groups (greens, cyanobacteria an diatoms), one zooplankton
group (cladocerans). This has the values in the default file and 25 \% parameter
ranges for sensitivity analysis.
}
\keyword{datasets}
