% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_pest.R
\name{have_pest}
\alias{have_pest}
\title{Is PEST++ available?}
\usage{
have_pest(dir = pest_install_dir())
}
\arguments{
\item{dir}{Character. Installation directory to search. Defaults to
\code{\link[=pest_install_dir]{pest_install_dir()}}.}
}
\value{
Logical.
}
\description{
Is PEST++ available?
}
