% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_fetch.R
\name{lt_filter}
\alias{lt_filter}
\title{Helper to construct filter parameters for lt_fetch}
\usage{
lt_filter(...)
}
\arguments{
\item{...}{R expressions representing filter conditions (e.g., \code{col == "value"}).}
}
\value{
A named list of filter conditions formatted for the API query.
}
\description{
This function takes R expressions like \code{col == "value"} and converts them
into the format expected by the API (e.g., \code{list(col = "eq.value")}). It
supports basic comparison operators (\code{==}, \code{!=}, \code{>}, \code{<}, \code{>=}, \code{<=}) and
\code{\%in\%}. It also handles \code{is.na(col)} and \code{!is.na(col)} for null checks.
}
\examples{
# Example usage of lt_filter - Get Lake Rotoehu contours with a specific 
# lernzmp_id
filters <- lt_filter(lernzmp_id == "LID40188")

}
