% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_fetch.R
\name{lt_schema}
\alias{lt_schema}
\title{Inspect the schema for a table in the Limnotrack database}
\usage{
lt_schema(table, base_url = "https://api.limnotrack.com/postgrest")
}
\arguments{
\item{table}{character; table name}

\item{base_url}{character; base URL of the PostgREST API}
}
\value{
a tibble with column names, types, formats and descriptions
}
\description{
Inspect the schema for a table in the Limnotrack database
}
