% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_fetch.R
\name{lt_tables}
\alias{lt_tables}
\title{List available tables in the Limnotrack database}
\usage{
lt_tables(base_url = "https://api.limnotrack.com/postgrest")
}
\arguments{
\item{base_url}{character; base URL of the PostgREST API}
}
\value{
a character vector of table names
}
\description{
List available tables in the Limnotrack database
}
