% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nz_aerial_imagery_metadata}
\alias{nz_aerial_imagery_metadata}
\title{sf object of aerial imagery metadata for New Zealand.}
\format{
\subsection{\code{nz_aerial_imagery_metadata}}{

A sf data frame with 230 rows and 8 columns:
\describe{
\item{layer_id}{Layer ID of the aerial imagery}
\item{title}{Title of the aerial imagery}
\item{abstract}{Abstract of the aerial imagery}
\item{region}{Region of the aerial imagery}
\item{res}{Resolution of the aerial imagery. Units are in denoted in the units column.}
\item{units}{Units of the resolution aerial imagery}
\item{year}{Year of the aerial imagery}
\item{geometry}{Geometry of the aerial imagery}
}
}
}
\source{
Package development.
}
\usage{
nz_aerial_imagery_metadata
}
\description{
This can be used to query aerial imagery coverage and metadata for New
Zealand.
}
\keyword{datasets}
