% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nz_dem_metadata}
\alias{nz_dem_metadata}
\title{sf object of Digital Elevation Model (DEM) metadata for New Zealand.}
\format{
\subsection{\code{nz_dem_metadata}}{

A sf data frame with 230 rows and 8 columns:
\describe{
\item{layer_id}{Layer ID of the DEM}
\item{title}{Title of the DEM}
\item{abstract}{Abstract of the DEM}
\item{region}{Region of the DEM}
\item{res}{Resolution of the DEM. Units are in denoted in the units column.}
\item{units}{Units of the resolution DEM}
\item{year}{Year of the DEM}
\item{geometry}{Geometry of the DEM}
}
}
}
\source{
Package development.
}
\usage{
nz_dem_metadata
}
\description{
This can be used to query DEM coverage and metadata for New Zealand.
}
\keyword{datasets}
