% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{param_var_matrix}
\alias{param_var_matrix}
\title{AEME parameters-response matrix}
\format{
\subsection{\code{param_var_matrix}}{

A data frame with 32 rows and 8 columns:#'
\describe{
\item{model}{The model that the parameter belongs to.}
\item{file}{The file that the parameter belongs to.}
\item{parameter_name}{The name of the parameter encoded as "group/name[index]".}
\item{HYD_temp}{TRUE if the parameter is associated with the HYD_temp variable, FALSE otherwise.}
\item{HYD_thmcln}{TRUE if the parameter is associated with the HYD_thmcln variable, FALSE otherwise.}
\item{LKE_lvlwtr}{TRUE if the parameter is associated with the LKE_lvlwtr variable, FALSE otherwise.}
\item{CHM_oxy}{TRUE if the parameter is associated with the CHM_oxy variable, FALSE otherwise.}
\item{PHY_tchla}{TRUE if the parameter is associated with the PHY_tchla variable, FALSE otherwise.}
}
}
}
\source{
Package development.
}
\usage{
param_var_matrix
}
\description{
A data frame that contains the parameters and their associated variables for the AEME model.
It contains a TRUE/FALSE value for each parameter and variable combination,
indicating whether the parameter is associated with the variable or not.
It can be used within calibration to determine which parameters are
associated with which variables.
}
\keyword{datasets}
