% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_pest.R
\name{pest_executables}
\alias{pest_executables}
\title{Names of the PEST++ executables aemetools knows about}
\usage{
pest_executables()
}
\value{
Character vector of executable base names (no extension).
}
\description{
Names of the PEST++ executables aemetools knows about
}
