% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_calib_summary.R
\name{plot_calib_summary}
\alias{plot_calib_summary}
\title{Plot calibration results}
\usage{
plot_calib_summary(calib, fit_col, nrow = 2, base_size = 8, log_y = TRUE)
}
\arguments{
\item{calib}{dataframe; output from \code{\link{read_calib}}}

\item{fit_col}{character; name of column containing fit values. If missing,
all fit types in \code{calib$simulation_data$fit_type} are used.}

\item{nrow}{integer; number of rows in plot}

\item{base_size}{numeric; base size for theme}

\item{log_y}{logical; use log scale on y-axis. Default is \code{TRUE}.}
}
\value{
list of plots
}
\description{
Plot calibration results
}
