% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multiscatter.R
\name{plot_multiscatter}
\alias{plot_multiscatter}
\title{Plot multiscatter of the parameters against each other}
\usage{
plot_multiscatter(sa)
}
\arguments{
\item{sa}{list; of sensitivity analysis results read in with \code{\link{read_sa}}}
}
\value{
list of \code{ggplot} objects
}
\description{
Plot multiscatter of the parameters against each other
}
