% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sobol.R
\name{plot_sobol}
\alias{plot_sobol}
\title{Visualization of first, total, second, third and fourth-order Sobol' indices}
\usage{
plot_sobol(sa)
}
\arguments{
\item{sa}{list; of sensitivity analysis results read in with \code{\link{read_sa}}}
}
\value{
list of \code{ggplot} objects
}
\description{
Visualization of first, total, second, third and fourth-order Sobol' indices
}
