% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-calib_helpers.R
\name{print.calib_sa_control}
\alias{print.calib_sa_control}
\title{Print a calibration or sensitivity analysis control object}
\usage{
\method{print}{calib_sa_control}(x, ...)
}
\arguments{
\item{x}{a \code{calib_sa_control} object, as created by
\code{\link{create_calib_control}} or \code{\link{create_sa_control}}.}

\item{...}{further arguments passed to or from other methods (unused).}
}
\value{
\code{x}, invisibly.
}
\description{
Print a calibration or sensitivity analysis control object
}
