% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_elev.R
\name{query_elev}
\alias{query_elev}
\title{Query elevation data for a shape within a DEM raster}
\usage{
query_elev(x, dem, lat, lon, layer_id = NULL)
}
\arguments{
\item{x}{sf object}

\item{dem}{SpatRaster object}

\item{lat}{numeric; latitude to query}

\item{lon}{numeric; longitude to query}

\item{layer_id}{integer; layer id of the DEM raster}
}
\value{
numeric; elevation value
}
\description{
This function queries the elevation data for a shape within a DEM raster. If
the latitude and longitude are not provided, the function will use the
centroid of the shape. If the elevation data is not found for the first
point, the function will try other points on the shape.
}
