% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_calib_meta.R
\name{read_calib_meta}
\alias{read_calib_meta}
\title{Read the calibration metadata}
\usage{
read_calib_meta(file, path = ".")
}
\arguments{
\item{file}{The path to the output file. It can either be a DuckDB database
or a CSV file.}

\item{path}{filepath; where input files are located relative to `config`.}
}
\value{
A data frame with the calibration metadata
}
\description{
This function reads the calibration metadata from the output file. The output
file can be either a DuckDB database or a CSV file. The function reads the
metadata and returns it as a data frame.
}
