% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_simulation_meta.R
\name{read_simulation_meta}
\alias{read_simulation_meta}
\title{Read the simulation metadata}
\usage{
read_simulation_meta(ctrl)
}
\arguments{
\item{ctrl}{A control object created by \code{\link{create_control}} and used
to run a calibration or a sensitivity analysis.}
}
\value{
A data frame with the simulation metadata
}
\description{
This function reads the simulation metadata from the output file. The output
file can be either a DuckDB database or a CSV file. The function reads the
metadata and returns it as a data frame.
}
