% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_simulation_output.R
\name{read_simulation_output}
\alias{read_simulation_output}
\alias{read_calib}
\title{Read calibration output}
\usage{
read_simulation_output(
  ctrl = NULL,
  file_name,
  file_dir,
  file_type = "db",
  sim_id = NULL,
  type
)

read_calib(
  ctrl = NULL,
  file_name,
  file_dir,
  file_type = "db",
  sim_id = NULL,
  type
)
}
\arguments{
\item{ctrl}{list; of controls for sensitivity analysis function created using
the \code{\link{create_control}} function. See \link{create_control} for
more details.}

\item{file_name}{The name of the output file. If \code{ctrl} is provided, this
argument is ignored.}

\item{file_dir}{The directory of the output file. If \code{ctrl} is provided, this
argument is ignored.}

\item{file_type}{string; file type to write the output to. Options are
\code{c("csv", "db")}. Defaults to "db".}

\item{sim_id}{A vector of simulation IDs to read. If NULL, all simulations
are read.}

\item{type}{A character string indicating the type of simulation. One of
"calib", "sa", or "all". If missing, the type is inferred from the
\code{ctrl} argument. If type is provided it overrides the \code{ctrl$method} value.}
}
\value{
A list with the metadata and simulation data frames.
}
\description{
Read calibration output
}
\examples{
aeme_file <- system.file("extdata/aeme.rds", package = "AEME")
aeme <- readRDS(aeme_file)
model_controls <- AEME::get_model_controls()
model <- c("glm_aed", "gotm_wet")
path <- "aeme"
aeme <- AEME::build_aeme(aeme = aeme, model = model, path = path,
                         model_controls = model_controls, ext_elev = 5) |>
  AEME::run_aeme()

data("aeme_parameters", package = "AEME")
param <- aeme_parameters

# Function to calculate fitness
nse <- function(df) {
# Calculate Nash-Sutcliffe Efficiency
  nse <- 1 - (sum((df$obs - df$model)^2) / sum((df$obs - mean(df$obs))^2))
  -1 * nse
}
FUN_list <- list(HYD_temp = nse, LKE_lvlwtr = nse)

ctrl <- create_control(method = "calib", NP = 10, itermax = 20, ncore = 2,
                       parallel = TRUE, file_type = "db",
                       file_name = "results.db")

vars_sim <- c("HYD_temp", "LKE_lvlwtr")
weights <- c("HYD_temp" = 1, "LKE_lvlwtr" = 1)

# Calibrate AEME model
sim_id <- calib_aeme(aeme = aeme, model = model, path = path,
                     param = param, FUN_list = FUN_list, ctrl = ctrl,
                     vars_sim = vars_sim, weights = weights)
                     
# Read calibration output                      
calib <- read_calib(sim_id = sim_id, ctrl = ctrl)
}
