% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_aeme_param.R
\name{run_aeme_param}
\alias{run_aeme_param}
\title{Run AEME with parameter dataframe}
\usage{
run_aeme_param(
  aeme,
  param,
  model,
  path = ".",
  model_controls = NULL,
  na_value = 999,
  return_nc = FALSE,
  return_aeme = FALSE,
  parallel = FALSE,
  verbose = FALSE,
  timeout = Inf
)
}
\arguments{
\item{aeme}{Aeme object.}

\item{param}{dataframe; of parameters read in from a csv file. Requires the
columns c("model", "file", "name", "value", "min", "max", "log")}

\item{model}{character vector; models to use. One or more of \code{"dy_cd"},
\code{"glm_aed"}, \code{"gotm_wet"}. Defaults to all models if not found in \code{aeme}.}

\item{path}{character; directory where input files are located. Defaults to
the path stored in \code{aeme}, or the current working directory if not set.}

\item{model_controls}{data.frame; model configuration, typically loaded via
\code{\link[AEME:get_model_controls]{get_model_controls()}}.}

\item{na_value}{numeric; value to return if model run is unsuccessful}

\item{return_nc}{boolean; return netCDF file connection}

\item{return_aeme}{boolean; return AEME object}

\item{parallel}{logical; run models in parallel. Defaults to FALSE.}

\item{verbose}{logical; print model output to console. Defaults to FALSE.}

\item{timeout}{Timeout for the process, in seconds, or as a \code{difftime}
object. If it is not finished before this, it will be killed.}
}
\value{
\code{na_value} if model run is unsuccessful
}
\description{
Run AEME with parameter dataframe
}
