% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_aeme_shiny.R
\name{run_aeme_shiny}
\alias{run_aeme_shiny}
\title{Run AEME in a Shiny app}
\usage{
run_aeme_shiny(aeme, param, path = ".", model_controls = NULL)
}
\arguments{
\item{aeme}{aeme; object.}

\item{param}{dataframe; of parameters read in from a csv file. Requires the
columns c("model", "file", "name", "value", "min", "max", "log")}

\item{path}{filepath; where input files are located relative to the current 
working directory.}

\item{model_controls}{dataframe; of configuration loaded from
"model_controls.csv".}
}
\value{
Launches shiny app
}
\description{
Run AEME in a Shiny app
}
