% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_GR.R
\name{run_GR}
\alias{run_GR}
\title{Run GR model}
\usage{
run_GR(
  hydro_model,
  param,
  warmup = NULL,
  run_index = NULL,
  IniStates = NULL,
  IniResLevels = NULL
)
}
\arguments{
\item{hydro_model}{\code{\linkS4class{HydroModel}}; object created by \code{\link{make_GR_inputs}()}.}

\item{param}{numeric vector; model parameters. Usually obtained from
\code{\link{calib_GR}()}.}

\item{warmup}{integer vector; row indices of \code{hydro_model@data} to use as
the warm-up period. Defaults to \code{1:(hydro_model@start - 1)}.}

\item{run_index}{integer vector; row indices of \code{hydro_model@data} to use
for the model run. Defaults to
\code{hydro_model@start:nrow(hydro_model@data)}.}

\item{IniStates}{(optional) [numeric] object of class \code{IniStates} [mm and °C], see \code{\link[airGR]{CreateIniStates}} for details}

\item{IniResLevels}{(optional) [numeric] vector of initial fillings for the GR stores (4 values; use NA when not relevant for a given model) [- and/or mm]. See details}
}
\value{
An \code{airGR} \code{OutputsModel} list with an additional
\code{catchment_area} element (m²).
}
\description{
Run a GR hydrological model from the \code{airGR} package using a \code{\linkS4class{HydroModel}}
object.
}
