% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_weights.R
\name{set_weights}
\alias{set_weights}
\title{Set weights for simulated variables}
\usage{
set_weights(vars_sim, weights = 1)
}
\arguments{
\item{vars_sim}{vector; of variables names to be used in the calculation of
model fit in \code{calib_aeme()} or \code{sa_aeme()}.}

\item{weights}{numeric vector; of weights for each variable in \code{vars_sim}.
If a single value is provided, it is recycled for all variables. Default is
1.}
}
\value{
named numeric vector; of weights for each variable in \code{vars_sim}.
}
\description{
Set weights for simulated variables
}
\examples{
vars_sim <- c("HYD_temp", "HYD_thmcln", "HYD_strat")
# Set same weight for all variables
weights1 <- set_weights(vars_sim)
print(weights1)
# Set different weights for each variable
weights2 <- set_weights(vars_sim, weights = c(1, 2, 0.5))
print(weights2)
}
