% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_csv_to_db.R
\name{write_csv_to_db}
\alias{write_csv_to_db}
\title{Write calibration/sensitivity analysis .csv files to database}
\usage{
write_csv_to_db(file_dir, file_name = "results.db")
}
\arguments{
\item{file_dir}{The directory where the .csv file is located.}

\item{file_name}{The name of the database file to write to. Default is
"results.db".}
}
\value{
file path to the database.
}
\description{
Write calibration/sensitivity analysis .csv files to database
}
