% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_aeme.R
\name{calib_aeme}
\alias{calib_aeme}
\title{Calibrate AEME model parameters using observations}
\usage{
calib_aeme(
  aeme,
  model,
  param,
  path,
  vars_sim = "HYD_temp",
  FUN_list,
  weights,
  model_controls = NULL,
  ctrl = NULL,
  param_var_matrix = NULL,
  param_df = NULL
)
}
\arguments{
\item{aeme}{aeme; object.}

\item{model}{vector; of models to be used. Can be `dy_cd`, `glm_aed`,
`gotm_wet`.}

\item{param}{dataframe; of parameters read in from a csv file. Requires the
columns c("model", "file", "name", "value", "min", "max", "log")}

\item{path}{filepath; where input files are located relative to the current 
working directory.}

\item{vars_sim}{vector; of variables names to be used in the calculation of
model fit.}

\item{FUN_list}{list of functions; named according to the variables in the
\code{vars_sim}. Funtions are of the form \verb{function(df)} which will be used
to calculate model fit. If nor provided, uses mean absolute error (MAE).}

\item{weights}{a named vector; of weights for each variable in vars_sim. If not
provided, defaults to 1 for each variable.}

\item{model_controls}{dataframe; of configuration loaded from
"model_controls.csv".}

\item{ctrl}{list; of controls for sensitivity analysis function created using
the \code{\link{create_control}} function. See \link{create_control} for
more details.}

\item{param_var_matrix}{list of dataframes; with parameters as rows and
response variables as columns. Created using
\code{\link{create_param_var_matrix}}. This is used to specify which
parameters are associated with which response variables, and therefore which
parameters are updated in each generation of the calibration.}

\item{param_df}{dataframe; of parameters to be used in the calibration.
Requires the columns c("model", "file", "name", "value", "min", "max"). This
is used to restart from a previous calibration.}
}
\value{
string of simulation id to be used to read the simulation output.
}
\description{
\code{calib_model()} runs the model and compares it against observations provided.
It can run in parallel by using multiple cores availlable on your computer
to run quicker.
}
\examples{
aeme_file <- system.file("extdata/aeme.rds", package = "AEME")
aeme <- readRDS(aeme_file)
model_controls <- AEME::get_model_controls()
model <- c("glm_aed", "gotm_wet")
aeme <- AEME::build_aeme(aeme = aeme, model = model, path = path, 
                         model_controls = model_controls,
                         ext_elev = 5, use_bgc = FALSE)
aeme <- AEME::run_aeme(aeme = aeme, model = model, path = path)

data("aeme_parameters", package = "AEME")
param <- aeme_parameters

# Function to calculate fitness
fit <- function(df) {
  mean(abs(df$obs - df$model))
}
FUN_list <- list(HYD_temp = fit, LKE_lvlwtr = fit)

ctrl <- create_control(method = "calib", NP = 10, itermax = 20, ncore = 2,
                       parallel = TRUE, file_type = "db",
                       file_name = "results.db")

vars_sim <- c("HYD_temp", "LKE_lvlwtr")
weights <- c("HYD_temp" = 1, "LKE_lvlwtr" = 1)

# Calibrate AEME model
sim_id <- calib_aeme(aeme = aeme, model = model, path = path,
                     param = param, FUN_list = FUN_list, ctrl = ctrl,
                     vars_sim = vars_sim, weights = weights)

}
