% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_aeme.R
\name{calib_aeme}
\alias{calib_aeme}
\title{Calibrate AEME model parameters using observations}
\usage{
calib_aeme(
  aeme,
  model,
  param,
  vars_sim = "HYD_temp",
  FUN_list,
  weights,
  path = ".",
  model_controls = NULL,
  ctrl = NULL,
  param_df = NULL
)
}
\arguments{
\item{aeme}{aeme; object.}

\item{model}{vector; of models to be used. Can be `dy_cd`, `glm_aed`,
`gotm_wet`.}

\item{param}{dataframe; of parameters read in from a csv file. Requires the
columns c("model", "file", "name", "value", "min", "max", "log")}

\item{vars_sim}{vector; of variables names to be used in the calculation of
model fit. Currently only supports using one variable.}

\item{FUN_list}{list of functions; named according to the variables in the
\code{vars_sim}. Funtions are of the form \verb{function(df)} which will be used
to calculate model fit. If nor provided, uses mean absolute error (MAE).}

\item{weights}{vector; of weights for each variable in vars_sim. If not
provided, defaults to 1 for each variable.}

\item{path}{filepath; where input files are located relative to the current 
working directory.}

\item{model_controls}{dataframe; of configuration loaded from
"model_controls.csv".}

\item{ctrl}{list; of controls for sensitivity analysis function created using
the \code{\link{create_control}} function. See \link{create_control} for
more details.}

\item{param_df}{dataframe; of parameters to be used in the calibration.
Requires the columns c("model", "file", "name", "value", "min", "max"). This
is used to restart from a previous calibration.}
}
\value{
string of simulation id to be used to read the simulation output.
}
\description{
\code{calib_model()} runs the model and compares it against observations provided.
It can run in parallel by using multiple cores availlable on your computer
to run quicker.
}
