% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_aeme.R
\name{calib_aeme}
\alias{calib_aeme}
\title{Calibrate AEME model parameters using observations}
\usage{
calib_aeme(
  aeme_data,
  path,
  param,
  model,
  mod_ctrls,
  vars_sim = "HYD_temp",
  FUN = NULL,
  ctrl = NULL,
  weights = c(1),
  param_df = NULL
)
}
\arguments{
\item{aeme_data}{aeme; data object.}

\item{path}{filepath; where input files are located relative to `config`.}

\item{param}{dataframe; of parameters read in from a csv file. Requires the
columns c("model", "file", "name", "value", "min", "max", "log")}

\item{model}{string; for which model. Options are c("dy_cd", "glm_aed" and
"gotm_wet")}

\item{mod_ctrls}{dataframe; of configuration loaded from
"mod_ctrls.csv".}

\item{vars_sim}{vector; of variables names to be used in the calculation of
model fit. Currently only supports using one variable.}

\item{FUN}{function; of the form \verb{function(O, P)} which will be used in
to calculate model fit. If NULL, uses mean absolute error (MAE).}

\item{ctrl}{list; of controls for calibration function.
\itemize{
\item \code{VTR} Value to be reached. The optimization process will stop if
either the maximum number of iterations itermax is reached or the best
parameter vector bestmem has found a value fn(bestmem) <= VTR. Default to
-Inf.
\item \code{NP} number of population members. Defaults to NA; if the user does not
change the value of NP from NA it is reset as
\code{10 * sum(param$model == model)}. For many  problems it is best to set NP
to be at least 10 times the length of the parameter vector.
\item \code{itermax} the maximum iteration (population generation) allowed.
Default is 200.
\item \code{reltol} relative convergence tolerance. The algorithm stops if it is
unable to reduce the value by a factor of \code{reltol * (abs(val) + reltol)}.
Default = 0.07
\item \code{p}: The quantile used to select the parents for the next generation. For
example, if \code{p = 0.25}, the best 25\% of the population will be used as
parents for the next generation.
\item \code{mutate} fraction of population to undergo mutation (0-1).
\item \code{parallel} boolean; run calibration in parallel. Default to TRUE
\item \code{out_file} filepath; to csv for calibration output to be written to.
Defaults to "results.csv"
\item \code{na_value} value to replace NA values with in observations. Default to
}
\enumerate{
\item 
}}

\item{weights}{vector; of weights for each variable in vars_sim. Default to
c(1).}
}
\value{
list; ctrl which was supplied with updated arguments if missing.
}
\description{
\code{calib_model()} runs the model and compares it against observations provided.
It can run in parallel by using multiple cores availlable on your computer
to run quicker.
}
