% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_GR.R
\name{calib_GR}
\alias{calib_GR}
\title{Calibrate GR model}
\usage{
calib_GR(
  inputs,
  warmup = NULL,
  run_index,
  FUN_CRIT = airGR::ErrorCrit_NSE,
  FUN_CALIB = airGR::Calibration_Michel,
  IniStates = NULL,
  IniResLevels = NULL
)
}
\arguments{
\item{inputs}{list; of inputs generated by \code{make_GR_inputs()}.}

\item{warmup}{vector; index of \code{inputs$data} to be used as a warmup period
for the model.}

\item{run_index}{vector; index of \code{inputs$data} to be used for running the
GR model.}

\item{FUN_CRIT}{[function (atomic or list)] error criterion function (e.g. \code{\link[airGR]{ErrorCrit_RMSE}}, \code{\link[airGR]{ErrorCrit_NSE}})}

\item{FUN_CALIB}{(optional) [function] calibration algorithm function (e.g. Calibration_Michel), default = \code{Calibration_Michel}}

\item{IniStates}{(optional) [numeric] object of class \code{IniStates} [mm and °C], see \code{\link[airGR]{CreateIniStates}} for details}

\item{IniResLevels}{(optional) [numeric] vector of initial fillings for the GR stores (4 values; use NA when not relevant for a given model) [- and/or mm]. See details}
}
\value{
list of airGR calibration outputs.
}
\description{
Calibrate GR model
}
