% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_era5_ncdf.R
\name{convert_era5_ncdf}
\alias{convert_era5_ncdf}
\title{Convert ERA5 netCDF files to AEME or LER}
\usage{
convert_era5_ncdf(
  lat,
  lon,
  variable = c("10m_u_component_of_wind", "10m_v_component_of_wind",
    "2m_dewpoint_temperature", "2m_temperature", "snowfall", "surface_pressure",
    "surface_solar_radiation_downwards", "surface_thermal_radiation_downwards",
    "total_precipitation"),
  year = 2022,
  site = "test",
  path = ".",
  format = "AEME"
)
}
\arguments{
\item{lat}{numeric; latitude}

\item{lon}{numeric; longitude}

\item{variable}{string with ERA5 variable names e.g. "2m_temperature", "total_precipitation"}

\item{year}{numeric; vector with years}

\item{site}{string of site name which was used when downloading the data.}

\item{path}{filepath to where the downloaded ERA5 ncdf files are stored.}

\item{format}{string; Either "AEME" or "LER". Default is "AEME".}
}
\description{
Convert ERA5 netCDF files to AEME or LakeEnsemblR formats.
}
