% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_control.R
\name{create_control}
\alias{create_control}
\title{Create control list (superseded)}
\usage{
create_control(method = c("calib", "sa"), ...)
}
\arguments{
\item{method}{Character. Either \code{"calib"} or \code{"sa"}.}

\item{...}{Arguments passed to the appropriate function.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{create_control()} is superseded. Please use
\code{\link[=create_calib_control]{create_calib_control()}} or \code{\link[=create_sa_control]{create_sa_control()}}
instead.
}
