% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_control.R
\name{create_control}
\alias{create_control}
\title{Create control list for calibration or sensitivity analysis}
\usage{
create_control(method, ...)
}
\arguments{
\item{method}{The method to be used. It can be either "calib" for calibration
or "sa" for sensitivity analysis.}

\item{...}{Additional arguments to be passed to the function
\code{create_control}. The arguments are different for calibration and
sensitivity analysis. There are arguments which are common to both methods:
\itemize{
\item \code{file_type} string; file type to write the output to. Options are c("csv",
"db"). Defaults to "db".
\item \code{file_name} string; file name to write the output to. Only used if \code{file_type}
is "db". Defaults to "results.db"
\item \code{na_value} value to replace NA values with in observations. Defaults to 999.
\item \code{parallel} boolean; run calibration in parallel. Default to TRUE
\item \code{ncore}: The number of cores to use for the calibration. This is only used
if \code{parallel = TRUE}. Default to \code{parallel::detectCores() - 1}.
}

For calibration, the arguments are:
\itemize{
\item \code{VTR} Value to be reached. The optimization process will stop if
either the maximum number of iterations itermax is reached or the best
parameter vector bestmem has found a value fn(bestmem) <= VTR. Default to
-Inf.
\item \code{NP} number of population members. Defaults to NA; if the user does not
change the value of NP from NA it is reset as
\code{10 * sum(param$model == model)}. For many  problems it is best to set NP
to be at least 10 times the length of the parameter vector.
\item \code{itermax} the maximum iteration (population generation) allowed.
Default is 200.
\item \code{reltol} relative convergence tolerance. The algorithm stops if it is
unable to reduce the value by a factor of \code{reltol * (abs(val) + reltol)}.
Default = 0.07
\item \code{cutoff}: The quantile cutoff used to select the parents for the next
generation. For example, if \code{cutoff = 0.25}, the best 25\% of the population
will be used as parents for the next generation.
\item \code{mutate} fraction of population to undergo mutation (0-1).
}

For sensitivity analysis, the arguments are:
\itemize{
\item \code{N}: The initial sample size of the base sample matrix.
\item \code{vars_sim}: A named list of output variables for sensitivity analysis.
The name is user defined but each list must contain:
\itemize{
\item \code{var}: The variable name to use for the sensitivity analysis.
\item \code{month}: A vector of months to use for the sensitivity analysis.
\item \code{depth_range}: A vector of length 2 with the minimum and maximum depth
range to use for the sensitivity analysis.
}
}}
}
\value{
list with the control parameters
}
\description{
Create control list for calibration or sensitivity analysis
}
\examples{
calib_ctrl <- create_control("calib", VTR = -Inf, NP = 10, itermax = 30,
                              reltol = 0.07, cutoff = 0.25, mutate = 0.1,
                              parallel = TRUE, file_type = "db",
                              file_name = "results.db", na_value = 999,
                              ncore = 3)

sa_ctrl <- create_control(method = "sa", N = 2^2, ncore = 2L, na_value = 999,
                          parallel = TRUE, file_type = "db",
                          file_name = "results.db",
                          vars_sim = list(surf_temp = list(var = "HYD_temp",
                                                           month = c(10:12, 1:3),
                                                           depth_range = c(0, 2)
                                                           ),
                                          bot_temp = list(var = "HYD_temp",
                                                          month = c(10:12, 1:3),
                                                          depth_range = c(10, 13)
                                                          )
                                          )
                        )


}
