% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_control.R
\name{create_control}
\alias{create_control}
\title{Create control list for calibration or sensitivity analysis}
\usage{
create_control(method, ...)
}
\arguments{
\item{method}{The method to be used. It can be either "calib" for calibration
or "sa" for sensitivity analysis.}

\item{...}{Additional arguments to be passed to the function.}
}
\value{
list with the control parameters
}
\description{
Create control list for calibration or sensitivity analysis
}
\examples{
ctrl <- create_control("calib", VTR = -Inf, NP = 10, itermax = 30,
                        reltol = 0.07, cutoff = 0.25, mutate = 0.1,
                        parallel = TRUE, out_file = "results.db",
                        na_value = 999, ncore = 3)


}
