% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_param_var_matrix.R
\name{create_param_var_matrix}
\alias{create_param_var_matrix}
\title{Create a parameter-variable matrix for each model}
\usage{
create_param_var_matrix(param, vars_sim)
}
\arguments{
\item{param}{dataframe; of parameters read in from a csv file. Requires the
columns c("model", "file", "name", "value", "min", "max", "log")}

\item{vars_sim}{vector; of variables names to be used in the calculation of
model fit.}
}
\value{
A list of data frames, one for each model, with parameters as rows
and variables as columns. The values in the data frames indicate whether a
parameter is associated with a variable (TRUE) or not (FALSE).
}
\description{
Create a parameter-variable matrix for each model
}
\examples{
param <- aeme_parameters
vars_sim <- c("HYD_temp", "CHM_oxy", "PHY_tchla")
param_var_matrix <- create_param_var_matrix(param, vars_sim)
}
