% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_param_var_matrix.R
\name{create_param_var_matrix}
\alias{create_param_var_matrix}
\title{Create a parameter-variable matrix for each model}
\usage{
create_param_var_matrix(param, vars_sim)
}
\arguments{
\item{param}{dataframe; of parameters read in from a csv file. Requires the
columns c("model", "file", "name", "value", "min", "max", "log")}

\item{vars_sim}{vector; of variables names to be used in the calculation of
model fit.}
}
\value{
A data frame with columns for model, file, name_full, and one
column for each variable in vars_sim. The variable columns contain TRUE/FALSE
values indicating whether the parameter is associated with the variable.
}
\description{
Create a parameter-variable matrix for each model
}
\examples{
param <- aeme_parameters
vars_sim <- c("HYD_temp", "CHM_oxy", "PHY_tchla")
param_var_matrix <- create_param_var_matrix(param, vars_sim)
}
