% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_best_params.R
\name{get_best_params}
\alias{get_best_params}
\title{Get best parameter values from calibration results}
\usage{
get_best_params(calib, fit_col = "fit", na_value = NULL)
}
\arguments{
\item{calib}{A list with the calibration results loaded using
\code{\link{read_calib}}.}

\item{fit_col}{character; name of column containing fit values. Default is
\code{"fit"}.}

\item{na_value}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Numeric. Penalty value
substituted for \code{NA} fit values, this is no longer needed as NA values
are now written to simulation_data in output of calib_aeme() and sa_aeme().
The argument will be removed in a future version.}
}
\value{
A data frame with the parameter values.
}
\description{
Get best parameter values from calibration results
}
