% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_era5_isimip_point.R
\name{get_era5_isimip_point}
\alias{get_era5_isimip_point}
\title{Get ERA5 data from ISIMIP3a for a point location}
\usage{
get_era5_isimip_point(lon, lat, years, vars, download_path = tempdir())
}
\arguments{
\item{lon}{numeric; Longitude}

\item{lat}{numeric; Latitude}

\item{years}{numeric; vector of years in numeric form to be extracted.
Currently years are limited to 1980-2022.}

\item{vars}{vector; with AEME meteorological variable names to be
downloaded. Defaults to all available variables: c("MET_tmpair",
"MET_tmpdew", "MET_wnduvu", "MET_wnduvv", "MET_pprain", "MET_ppsnow",
"MET_prsttn", "MET_radswd").}

\item{download_path}{character; path to download the data. Default is
the temporary directory.}
}
\value{
A data frame with the requested variables
}
\description{
Get ERA5 data from ISIMIP3a for a point location
}
\examples{
\dontrun{
lon <- 13.064332
lat <- 52.380551
years <- 2015:2021
vars <- c("MET_tmpair", "MET_pprain")
get_era5_isimip_point(lon, lat, years, vars)
}
}
