% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_era5_land_point_nz.R
\name{get_era5_land_point_nz}
\alias{get_era5_land_point_nz}
\title{Get ERA5 data for a coordinate}
\usage{
get_era5_land_point_nz(
  lat,
  lon,
  years,
  vars = c("MET_tmpair", "MET_tmpdew", "MET_wnduvu", "MET_wnduvv", "MET_pprain",
    "MET_ppsnow", "MET_prsttn", "MET_radswd"),
  api_url = "http://170.64.143.18:80",
  api_key = NULL
)
}
\arguments{
\item{lat}{numeric; Latitude}

\item{lon}{numeric; Longitude}

\item{years}{numeric; vector of years in numeric form to be extracted.
Currently years are limited to 1980-2022.}

\item{vars}{vector; with AEME meteorological variable names to be
downloaded. Defaults to all available variables: c("MET_tmpair",
"MET_tmpdew", "MET_wnduvu", "MET_wnduvv", "MET_pprain", "MET_ppsnow",
"MET_prsttn", "MET_radswd").}

\item{api_url}{character; URL to the API endpoint. Default is
"http://170.64.143.18:80"}

\item{api_key}{character; API key to access the data.}
}
\value{
dataframe of daily ERA5 data.
}
\description{
Extract \href{https://www.ecmwf.int/en/era5-land}{ERA5-Land} meteorological data
for the closest grid to a particular latitude and longitude for a selected
number of years (1980-2023).
}
