% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_era5_point.R
\name{get_era5_point}
\alias{get_era5_point}
\title{Extract data from ERA5 RDS and format for `LakeEnsemblR```}
\usage{
get_era5_point(lat, lon, years, variables, format = "aeme", parallel = FALSE)
}
\arguments{
\item{lat}{numeric; Latitude}

\item{lon}{numeric; Longitude}

\item{years}{numeric; vector of years in numeric form to be extracted}

\item{variables}{vector; with lookup table for variable names}

\item{format}{string; the format for column headers. Can be "aeme" or "ler".
Defaults to "aeme".}
}
\value{
dataframe of daily ERA5 data.
}
\description{
Extract met data frame from closest grid to lat & lon
}
