% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_linz_basemap_tile.R
\name{get_linz_basemap_tile}
\alias{get_linz_basemap_tile}
\title{Get LINZ basemap tile}
\usage{
get_linz_basemap_tile(x, zoom = 16, key = NULL, verbose = FALSE)
}
\arguments{
\item{x}{sf, sfc, bbox, SpatRaster, SpatVector or SpatExtent object.
If \code{x} is a SpatExtent it
must express coordinates in lon/lat WGS84 (epsg:4326).}

\item{zoom}{zoom level (see Details).}

\item{key}{character; LINZ API key. If NULL, will look for the key in the
LINZ_KEY environment variable. If that is not set, will throw an error.
Use the \code{add_linz_key} function to set the key.

@importFrom maptiles create_provider get_tiles
@importFrom terra subset}

\item{verbose}{if TRUE, tiles filepaths, zoom level and
attribution are displayed.}
}
\value{
A SpatRaster is returned.
}
\description{
Get map tiles from the LINZ basemap service. This requires an API key which
can be obtained from: https://basemaps.linz.govt.nz
}
