% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param.R
\name{get_param}
\alias{get_param}
\title{Get parameter values from calibration results}
\usage{
get_param(calib, na_value, fit_col = "fit", best = FALSE, quantile = 0.1)
}
\arguments{
\item{calib}{A list with the calibration results loaded using
\code{\link{read_calib}}.}

\item{na_value}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Numeric. Penalty value
substituted for \code{NA} fit values, this is no longer needed as NA values
are now written to simulation_data in output of calib_aeme() and sa_aeme().
The argument will be removed in a future version.}

\item{fit_col}{character; name of column containing fit values. Default is
\code{"fit"}.}

\item{best}{A logical value indicating whether to return the best parameter
values or all parameter values.}

\item{quantile}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} The quantile to use for
the top quantile of the fit_value. Defaults to 0.1. This is no longer needed
and will be removed in a future version.}
}
\value{
A data frame with the parameter values.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{get_param()} is deprecated. Please use either \link{get_sim_params()} or
\link{get_best_params()} instead, depending on whether you want to retrieve
all parameter values or just the best parameter values based on a specified
fit column and quantile threshold.
}
