% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param.R
\name{get_param}
\alias{get_param}
\title{Get parameter values from calibration results}
\usage{
get_param(calib, na_value, fit_col = "fit", best = FALSE)
}
\arguments{
\item{calib}{A list with the calibration results loaded using
\code{\link{read_calib}}.}

\item{na_value}{A numeric value which corresponds to the NA value used in
the calibration.}

\item{fit_col}{character; name of column containing fit values. Default is
\code{"fit"}.}

\item{best}{A logical value indicating whether to return the best parameter
values or all parameter values.}
}
\value{
A data frame with the parameter values.
}
\description{
Get parameter values from calibration results
}
