% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param.R
\name{get_param}
\alias{get_param}
\title{Get parameter values from calibration results}
\usage{
get_param(calib, model, na_value, best = FALSE)
}
\arguments{
\item{calib}{A data frame with the calibration results.}

\item{model}{A character vector with the model name.}

\item{na_value}{A numeric value which corresponds to the NA value used in
the calibration.}

\item{best}{A logical value indicating whether to return the best parameter
values or all parameter values.}
}
\value{
A data frame with the parameter values.
}
\description{
Get parameter values from calibration results
}
