% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raster_tile.R
\name{get_raster_tile}
\alias{get_raster_tile}
\title{Get a raster tile from LINZ}
\usage{
get_raster_tile(shape, layer_id, zoom = 15, key = NULL, verbose = FALSE)
}
\arguments{
\item{shape}{sf object with a single point or polygon.}

\item{layer_id}{numeric; layer ID value for the raster layer on the LINZ
data service. See \url{https://data.linz.govt.nz/}}

\item{zoom}{numeric; zoom level for the tile. Default is 15.}

\item{key}{character; LINZ API key. If NULL, will look for the key in the
LINZ_KEY environment variable. If that is not set, will throw an error.
Use the \code{add_linz_key} function to set the key.

@importFrom maptiles create_provider get_tiles
@importFrom terra subset}
}
\value{
a raster object
}
\description{
Get a raster tile from LINZ
}
