% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_simulation_time.R
\name{get_simulation_time}
\alias{get_simulation_time}
\title{Get simulation time for each model}
\usage{
get_simulation_time(aeme, model, path, param, FUN_list, vars_sim, weights)
}
\arguments{
\item{aeme}{aeme; object.}

\item{model}{vector; of models to be used. Can be `dy_cd`, `glm_aed`,
`gotm_wet`.}

\item{path}{filepath; where input files are located relative to the current 
working directory.}

\item{param}{dataframe; of parameters read in from a csv file. Requires the
columns c("model", "file", "name", "value", "min", "max", "log")}

\item{FUN_list}{list of functions; named according to the variables in the
\code{vars_sim}. Funtions are of the form \verb{function(df)} which will be used
to calculate model fit. If nor provided, uses mean absolute error (MAE).}

\item{vars_sim}{vector; of variables names to be used in the calculation of
model fit. Currently only supports using one variable.}

\item{weights}{vector; of weights for each variable in vars_sim. If not
provided, defaults to 1 for each variable.}
}
\value{
numeric vector of simulation time for each model.
}
\description{
Get simulation time for each model
}
\examples{
tmpdir <- tempdir()
aeme_dir <- system.file("extdata/lake/", package = "AEME")
# Copy files from package into tempdir
file.copy(aeme_dir, tmpdir, recursive = TRUE)
path <- file.path(tmpdir, "lake")
aeme <- AEME::yaml_to_aeme(path = path, "aeme.yaml")
model_controls <- AEME::get_model_controls()
model <- c("glm_aed", "gotm_wet")
aeme <- AEME::build_aeme(path = path, aeme = aeme,
model = model, model_controls = model_controls,
ext_elev = 5, use_bgc = FALSE)
aeme <- AEME::run_aeme(aeme = aeme, model = model, path = path)
data("aeme_parameters", package = "AEME")
param <- aeme_parameters
# Function to calculate fitness
fit <- function(df) {
mean(abs(df$obs - df$model))
}
FUN_list <- list(HYD_temp = fit, LKE_lvlwtr = fit)
ctrl <- create_control(method = "calib", NP = 10, itermax = 30, ncore = 2,
parallel = TRUE, file_type = "db", file_name = "results.db")
vars_sim <- c("HYD_temp", "LKE_lvlwtr")
weights <- c("HYD_temp" = 10, "LKE_lvlwtr" = 1)
sim_times <- get_simulation_time(aeme = aeme, model = model, path = path,
param = param, FUN_list = FUN_list, vars_sim = vars_sim, weights = weights)
}
