% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_calib.R
\name{plot_calib}
\alias{plot_calib}
\title{Plot calibration results}
\usage{
plot_calib(
  calib,
  na_value,
  fit_col = "fit",
  nrow = 2,
  base_size = 8,
  return_pars = FALSE,
  log_y = TRUE
)
}
\arguments{
\item{calib}{dataframe; output from \code{\link{read_calib}}}

\item{na_value}{A numeric value which corresponds to the NA value used in
the calibration.}

\item{fit_col}{character; name of column containing fit values. Default is
\code{"fit"}.}

\item{nrow}{integer; number of rows in plot}

\item{base_size}{numeric; base size for theme}

\item{return_pars}{logical; return parameter values}

\item{log_y}{logical; use log scale on y-axis. Default is \code{TRUE}.}
}
\value{
list of plots
}
\description{
Plot calibration results
}
\examples{
aeme_file <- system.file("extdata/aeme.rds", package = "AEME")
aeme <- readRDS(aeme_file)
model_controls <- AEME::get_model_controls()
model <- c("glm_aed", "gotm_wet")
path <- "aeme"
aeme <- AEME::build_aeme(aeme = aeme, model = model, path = path,
                         model_controls = model_controls, ext_elev = 5) |>
  AEME::run_aeme()

data("aeme_parameters", package = "AEME")
param <- aeme_parameters

# Function to calculate fitness
nse <- function(df) {
# Calculate Nash-Sutcliffe Efficiency
  nse <- 1 - (sum((df$obs - df$model)^2) / sum((df$obs - mean(df$obs))^2))
  -1 * nse
}
FUN_list <- list(HYD_temp = nse, LKE_lvlwtr = nse)

ctrl <- create_control(method = "calib", NP = 10, itermax = 20, ncore = 2,
                       parallel = TRUE, file_type = "db",
                       file_name = "results.db")

vars_sim <- c("HYD_temp", "LKE_lvlwtr")
weights <- c("HYD_temp" = 1, "LKE_lvlwtr" = 1)

# Calibrate AEME model
sim_id <- calib_aeme(aeme = aeme, model = model, path = path,
                     param = param, FUN_list = FUN_list, ctrl = ctrl,
                     vars_sim = vars_sim, weights = weights)
                     
# Read calibration output                      
calib <- read_calib(sim_id = sim_id, ctrl = ctrl)
plist <- plot_calib(calib = calib)

# Dotty plot
plist$dotty

# Convergence plot
plist$convergence

# Histogram plot
plist$histogram
}
