% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_calib.R
\name{plot_calib}
\alias{plot_calib}
\title{Plot calibration results}
\usage{
plot_calib(
  calib,
  na_value,
  fit_col = "fit",
  nrow = 2,
  base_size = 8,
  return_pars = FALSE,
  log_y = TRUE
)
}
\arguments{
\item{calib}{dataframe; output from \code{\link{read_calib}}}

\item{na_value}{numeric; value to replace NA values with}

\item{fit_col}{character; name of column containing fit values. Default is
\code{"fit"}.}

\item{nrow}{integer; number of rows in plot}

\item{base_size}{numeric; base size for theme}

\item{return_pars}{logical; return parameter values}

\item{log_y}{logical; use log scale on y-axis. Default is \code{TRUE}.}
}
\value{
list of plots
}
\description{
Plot calibration results
}
