% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_calib.R
\name{plot_calib}
\alias{plot_calib}
\title{Plot calibration results}
\usage{
plot_calib(
  calib,
  model,
  na_value,
  nrow = 2,
  base_size = 8,
  return_pars = FALSE
)
}
\arguments{
\item{calib}{dataframe; output from \code{\link{read_calib}}}

\item{model}{string; for which model. Options are c("dy_cd", "glm_aed" and
"gotm_wet")}

\item{na_value}{numeric; value to replace NA values with}

\item{nrow}{integer; number of rows in plot}

\item{base_size}{numeric; base size for theme}

\item{return_pars}{logical; return parameter values}
}
\value{
list of plots
}
\description{
Plot calibration results
}
