% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ensemble.R
\name{plot_ensemble}
\alias{plot_ensemble}
\title{Plot AEME ensemble output}
\usage{
plot_ensemble(
  aeme,
  model,
  var_sim = "HYD_temp",
  depth = NULL,
  conf_int = 0.95,
  type = "ribbon",
  remove_spin_up = TRUE,
  add_obs = TRUE,
  var_lims = NULL
)
}
\arguments{
\item{aeme}{aeme; object.}

\item{model}{vector; of models to be used. Can be `dy_cd`, `glm_aed`,
`gotm_wet`.}

\item{var_sim}{string; of variable to plot}

\item{depth}{numeric; depth of the variable to extract. Default is NULL. If
NULL, the variable profiles are extracted.}

\item{conf_int}{numeric; confidence interval to plot when \code{type = "ribbon"}.
Default is 0.95.}

\item{type}{character; type of plot to create. Can be \code{"ribbon"}, where it
plots a geom_ribbon to represent the confidence intervals specified in
\code{conf_int} or \verb{"line}, where it plots all the ensemble members as lines.
Default is "ribbon".}

\item{remove_spin_up}{logical; remove spin-up period from plot. Default is
TRUE.}

\item{add_obs}{logical; add observations to plot}

\item{var_lims}{numeric vector of length 2; limits for the variable.
Defaults to NULL and will generate common limits for all variables.}
}
\value{
ggplot object
}
\description{
Plot AEME ensemble output
}
