% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Plot the scatter of the output of a sensitivity analysis}
\usage{
plot_scatter(sa)
}
\arguments{
\item{sa}{list; of sensitivity analysis results read in with \code{\link{read_sa}}}
}
\value{
\code{ggplot} object
}
\description{
Plot the scatter of the output of a sensitivity analysis
}
