% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sa.R
\name{read_sa}
\alias{read_sa}
\title{Read sensitivity analysis output}
\usage{
read_sa(ctrl, sim_id, R = NULL, boot = TRUE)
}
\arguments{
\item{ctrl}{list; of controls for sensitivity analysis function created using
the \code{\link{create_control}} function. See \link{create_control} for
more details.}

\item{R}{Positive integer, number of bootstrap replicas. Default is NULL.}

\item{boot}{Logical. If TRUE, the function bootstraps the Sobol' indices. If FALSE, it provides point
estimates. Default is \code{boot = FALSE}.}
}
\value{
A list with thedata frame with the sensitivity analysis results and
the sobol indices for each variable.
}
\description{
Read sensitivity analysis output
}
