% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_simulation_meta.R
\name{read_simulation_meta}
\alias{read_simulation_meta}
\title{Read the simulation metadata}
\usage{
read_simulation_meta(ctrl = NULL, file_name, file_dir, type)
}
\arguments{
\item{ctrl}{A list with the control parameters used in \code{calib_aeme()} or
\code{sa_aeme()}. If \code{ctrl} is provided, the \code{file_name} and \code{file_dir} arguments
are ignored.}

\item{file_name}{The name of the output file. If \code{ctrl} is provided, this
argument is ignored.}

\item{file_dir}{The directory of the output file. If \code{ctrl} is provided, this
argument is ignored.}

\item{type}{Optional character vector to filter by simulation type. Possible
values are "sa" for sensitivity analysis simulations and "calib" for
calibration simulations. If not provided, all simulations are returned.}
}
\value{
A data frame with the simulation metadata
}
\description{
This function reads the simulation metadata from the output file. The output
file can be either a DuckDB database or a CSV file. The function reads the
metadata and returns it as a data frame.
}
