% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_simulation_output.R
\name{read_simulation_output}
\alias{read_simulation_output}
\alias{read_calib}
\title{Read calibration output}
\usage{
read_simulation_output(ctrl, path = ".", file = NULL, sim_id = NULL)

read_calib(ctrl, path = ".", file = NULL, sim_id = NULL)
}
\arguments{
\item{ctrl}{list; of controls for sensitivity analysis function created using
the \code{\link{create_control}} function. See \link{create_control} for
more details.}

\item{path}{filepath; where input files are located relative to `config`.}

\item{file}{file name; to read. Can be either a "csv" or "db" file and it
must be located within the path.}

\item{sim_id}{A vector of simulation IDs to read.}
}
\value{
A list with the metadata and simulation data frames.
}
\description{
Read calibration output
}
