% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_simulation_output.R
\name{read_simulation_output}
\alias{read_simulation_output}
\alias{read_calib}
\title{Read calibration output}
\usage{
read_simulation_output(ctrl = NULL, file_name, file_dir, sim_id = NULL)

read_calib(ctrl = NULL, file_name, file_dir, sim_id = NULL)
}
\arguments{
\item{ctrl}{list; of controls for sensitivity analysis function created using
the \code{\link{create_control}} function. See \link{create_control} for
more details.}

\item{file_name}{The name of the output file. If \code{ctrl} is provided, this
argument is ignored.}

\item{file_dir}{The directory of the output file. If \code{ctrl} is provided, this
argument is ignored.}

\item{sim_id}{A vector of simulation IDs to read.}
}
\value{
A list with the metadata and simulation data frames.
}
\description{
Read calibration output
}
