% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_simulation_output.R
\name{read_simulation_output}
\alias{read_simulation_output}
\alias{read_calib}
\title{Read calibration output}
\usage{
read_simulation_output(ctrl, path = ".", file = NULL, sim_id = NULL)

read_calib(ctrl, path = ".", file = NULL, sim_id = NULL)
}
\arguments{
\item{ctrl}{list; of controls for sensitivity analysis function created using
the \code{\link{create_control}} function. See \link{create_control} for
more details.}

\item{path}{filepath; where input files are located relative to `config`.}

\item{raw}{logical. If TRUE, return the raw calibration output as a dataframe
with the "fit" and "gen" columns. This is generally used when restarting a
calibration.}
}
\value{
A data frame with the calibration results.
}
\description{
Read calibration output
}
