% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_aeme_ensemble.R
\name{run_aeme_ensemble}
\alias{run_aeme_ensemble}
\title{Run AEME ensemble}
\usage{
run_aeme_ensemble(
  aeme,
  model,
  n = 10,
  dist = c("norm", "unif"),
  path = ".",
  parallel = FALSE,
  ncore = NULL,
  param = NULL,
  na_value = 999
)
}
\arguments{
\item{aeme}{Aeme object.}

\item{model}{character vector; models to use. One or more of \code{"dy_cd"},
\code{"glm_aed"}, \code{"gotm_wet"}. Defaults to all models if not found in \code{aeme}.}

\item{n}{numeric; number of ensemble members to generate.}

\item{dist}{character; distribution to sample from. Default is "norm". Other
options are "unif".}

\item{path}{character; directory where input files are located. Defaults to
the path stored in \code{aeme}, or the current working directory if not set.}

\item{parallel}{logical; whether to run in parallel. Default is FALSE.}

\item{ncore}{numeric; number of cores to use for parallel processing. Default
is NULL. If NULL, the function will use the number of cores minus 1.}

\item{param}{data.frame; parameter values to use. Default is NULL. If NULL,
the function will use the parameters from the aeme object.}

\item{na_value}{numeric; value to use for NA values. Default is 999.}
}
\value{
an \code{aeme} object with model output loaded.
}
\description{
Run a parameter ensemble
}
