% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_aeme_ensemble.R
\name{run_aeme_ensemble}
\alias{run_aeme_ensemble}
\title{Run AEME ensemble}
\usage{
run_aeme_ensemble(
  aeme,
  model,
  n = 10,
  dist = "norm",
  path = ".",
  parallel = FALSE,
  ncores = NULL,
  param = NULL,
  na_value = 999
)
}
\arguments{
\item{aeme}{aeme; object.}

\item{model}{string; for which model to calibrate. Only one model can be
passed. Options are c("dy_cd", "glm_aed" and "gotm_wet").}

\item{path}{filepath; where input files are located relative to `config`.}

\item{param}{dataframe; of parameters read in from a csv file. Requires the
columns c("model", "file", "name", "value", "min", "max", "log")}
}
\value{
an `aeme` object with model output loaded.
}
\description{
Run a parameter ensemble
}
