% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_aeme_ensemble.R
\name{run_aeme_ensemble}
\alias{run_aeme_ensemble}
\title{Run AEME ensemble}
\usage{
run_aeme_ensemble(
  aeme,
  model,
  n = 10,
  dist = "norm",
  path = ".",
  parallel = FALSE,
  ncores = NULL,
  param = NULL,
  na_value = 999
)
}
\arguments{
\item{aeme}{aeme; object.}

\item{model}{string; for which model to calibrate. Only one model can be
passed. Options are c("dy_cd", "glm_aed" and "gotm_wet").}

\item{n}{numeric; number of ensemble members to generate.}

\item{dist}{character; distribution to sample from. Default is "norm". Other
options are "unif".}

\item{path}{filepath; where input files are located relative to `config`.}

\item{parallel}{logical; whether to run in parallel. Default is FALSE.}

\item{ncores}{numeric; number of cores to use. Default is NULL.}

\item{param}{data.frame; parameter values to use. Default is NULL. If NULL,
the function will use the parameters from the aeme object.}

\item{na_value}{numeric; value to use for NA values. Default is 999.}
}
\value{
an `aeme` object with model output loaded.
}
\description{
Run a parameter ensemble
}
