% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_aeme_ensemble.R
\name{run_aeme_ensemble}
\alias{run_aeme_ensemble}
\title{Run AEME ensemble}
\usage{
run_aeme_ensemble(aeme, sim_id, param, calib = NULL, fit_col = "fit", ctrl)
}
\arguments{
\item{aeme}{aeme; object.}

\item{param}{dataframe; of parameters read in from a csv file. Requires the
columns c("model", "file", "name", "value", "min", "max", "log")}

\item{ctrl}{list; of controls for sensitivity analysis function created using
the \code{\link{create_control}} function. See \link{create_control} for
more details.}
}
\value{
an `aeme` object with model output loaded.
}
\description{
Run a parameter ensemble
}
