% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_aeme_ensemble.R
\name{run_aeme_ensemble}
\alias{run_aeme_ensemble}
\title{Run AEME ensemble}
\usage{
run_aeme_ensemble(
  aeme,
  model,
  n = 10,
  dist = "norm",
  path = ".",
  parallel = FALSE,
  ncore = NULL,
  param = NULL,
  na_value = 999
)
}
\arguments{
\item{aeme}{aeme; object.}

\item{model}{vector; of models to be used. Can be `dy_cd`, `glm_aed`,
`gotm_wet`.}

\item{n}{numeric; number of ensemble members to generate.}

\item{dist}{character; distribution to sample from. Default is "norm". Other
options are "unif".}

\item{path}{filepath; where input files are located relative to `config`.}

\item{parallel}{logical; whether to run in parallel. Default is FALSE.}

\item{ncore}{numeric; number of cores to use for parallel processing. Default
is NULL. If NULL, the function will use the number of cores minus 1.}

\item{param}{data.frame; parameter values to use. Default is NULL. If NULL,
the function will use the parameters from the aeme object.}

\item{na_value}{numeric; value to use for NA values. Default is 999.}
}
\value{
an `aeme` object with model output loaded.
}
\description{
Run a parameter ensemble
}
