% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_aeme_param.R
\name{run_aeme_param}
\alias{run_aeme_param}
\title{Run AEME with parameter dataframe}
\usage{
run_aeme_param(
  aeme_data,
  param,
  model,
  path,
  mod_ctrls,
  na_value = 999,
  return_nc = FALSE
)
}
\arguments{
\item{aeme_data}{aeme; data object.}

\item{param}{dataframe; of parameters read in from a csv file. Requires the
columns c("model", "file", "name", "value", "min", "max", "log")}

\item{model}{vector; of models to be used. Can be `dy_cd`, `glm_aed`,
`gotm_wet`.}

\item{path}{filepath; where input files are located relative to `config`.}

\item{mod_ctrls}{dataframe; of configuration loaded from
"mod_ctrls.csv".}

\item{na_value}{numeric; value to return if model run is unsuccessful}

\item{return_nc}{boolean; return netCDF file connection}
}
\value{
\code{na_value} if model run is unsuccessful
}
\description{
Run AEME with parameter dataframe
}
