% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_aeme_shiny.R
\name{run_aeme_shiny}
\alias{run_aeme_shiny}
\title{Run AEME in a Shiny app}
\usage{
run_aeme_shiny(aeme, param, path = ".", model_controls = NULL)
}
\arguments{
\item{aeme}{Aeme object.}

\item{param}{dataframe; of parameters read in from a csv file. Requires the
columns c("model", "file", "name", "value", "min", "max", "log")}

\item{path}{character; directory where input files are located. Defaults to
the path stored in \code{aeme}, or the current working directory if not set.}

\item{model_controls}{data.frame; model configuration, typically loaded via
\code{\link[AEME:get_model_controls]{get_model_controls()}}.}
}
\value{
Launches shiny app
}
\description{
Run AEME in a Shiny app
}
