% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_aeme_shiny.R
\name{run_aeme_shiny}
\alias{run_aeme_shiny}
\title{Run AEME in a Shiny app}
\usage{
run_aeme_shiny(aeme_data, param, path, mod_ctrls)
}
\arguments{
\item{aeme_data}{aeme; data object.}

\item{path}{filepath; where input files are located relative to `config`.}

\item{mod_ctrls}{dataframe; of configuration loaded from
"mod_ctrls.csv".}
}
\value{
Launches shiny app
}
\description{
Run AEME in a Shiny app
}
