% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_and_fit.R
\name{run_and_fit}
\alias{run_and_fit}
\title{Run a model and calculate model fit.}
\usage{
run_and_fit(
  aeme,
  param,
  model,
  vars_sim,
  path,
  model_controls = NULL,
  FUN_list = NULL,
  weights,
  na_value = 999,
  var_indices = NULL,
  return_indices = FALSE,
  include_wlev = FALSE,
  return_df = FALSE,
  method = "calib",
  sa_ctrl = NULL,
  fit = TRUE
)
}
\arguments{
\item{aeme}{aeme; object.}

\item{param}{dataframe; of parameters read in from a csv file. Requires the
columns c("model", "file", "name", "value", "min", "max", "log")}

\item{model}{string; for which model. Options are c("dy_cd", "glm_aed" and
"gotm_wet")}

\item{vars_sim}{vector; of variables names to be used in the calculation of
model fit. Currently only supports using one variable.}

\item{path}{filepath; where input files are located relative to `config`.}

\item{model_controls}{dataframe; of configuration loaded from
"model_controls.csv".}

\item{FUN_list}{function; of the form \verb{function(O, P)} which will be used in
to calculate model fit. If NULL, uses mean absolute error (MAE).}

\item{weights}{vector; of weights to be used in the calculation of model fit.}

\item{na_value}{numeric; value to be returned if model fails to run.}

\item{var_indices}{list; generated from running \code{run_and_fit()} with
\verb{return indices = TRUE} on the first simulation.}

\item{return_indices}{boolean; return the indices (depths, time and dates)
of each variable. Used when running calibration and the time period does not
change between simulations.}

\item{include_wlev}{boolean; include water level in the calculation of model
fit.}

\item{return_df}{boolean; return dataframe of modelled and observed.}

\item{method}{string; of the method of the model run. Options are c("sa",
"calib").}

\item{fit}{boolean; fit model or not. If FALSE, only return netCDF file
connection.}
}
\value{
A single value of model fit, calculated by \code{FUN_list}.
}
\description{
Run a model and calculate model fit.
}
