% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_GR.R
\name{run_GR}
\alias{run_GR}
\title{Run GR model}
\usage{
run_GR(
  inputs,
  param,
  warmup = NULL,
  run_index,
  IniStates = NULL,
  IniResLevels = NULL
)
}
\arguments{
\item{inputs}{list; of inputs generated by \code{make_GR_inputs()}.}

\item{param}{vector; of parameters. Usually generated from \code{calib_GR()}.}

\item{warmup}{vector; index of \code{inputs$data} to be used as a warmup period
for the model.}

\item{run_index}{vector; index of \code{inputs$data} to be used for running the
GR model.}

\item{IniStates}{(optional) [numeric] object of class \code{IniStates} [mm and °C], see \code{\link[airGR]{CreateIniStates}} for details}

\item{IniResLevels}{(optional) [numeric] vector of initial fillings for the GR stores (4 values; use NA when not relevant for a given model) [- and/or mm]. See details}
}
\value{
list of GR model output.
}
\description{
Run GR model
}
