% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_aeme.R
\name{sa_aeme}
\alias{sa_aeme}
\title{Run sensitivity analysis on AEME model parameters}
\usage{
sa_aeme(
  aeme_data,
  path = ".",
  param,
  model,
  mod_ctrls,
  FUN_list = NULL,
  ctrl = NULL,
  param_df = NULL
)
}
\arguments{
\item{aeme_data}{aeme; data object.}

\item{path}{filepath; where input files are located relative to `config`.}

\item{param}{dataframe; of parameters read in from a csv file. Requires the
columns c("model", "file", "name", "value", "min", "max", "log")}

\item{model}{string; for which model to calibrate. Only one model can be
passed. Options are c("dy_cd", "glm_aed" and "gotm_wet").}

\item{mod_ctrls}{dataframe; of configuration loaded from
"mod_ctrls.csv".}

\item{FUN_list}{list of functions; named according to the variables in the
\code{vars_sim}. Funtions are of the form \verb{function(df)} which will be used
to calculate model fit. If NULL, uses mean absolute error (MAE).}

\item{ctrl}{list; of controls for sensitivity analysis function created using
the \code{\link{create_control}} function. See \link{create_control} for
more details.}

\item{param_df}{dataframe; of parameters to be used in the calibration.
Requires the columns c("model", "file", "name", "value", "min", "max"). This
is used to restart from a previous calibration.}
}
\value{
string of simulation id to be used to read the simulation output.
}
\description{
\code{calib_model()} runs the model and compares it against observations provided.
It can run in parallel by using multiple cores availlable on your computer
to run quicker.
}
\examples{
\dontrun{
  # Run sensitivity analysis
  tmpdir <- tempdir()
  aeme_dir <- system.file("extdata/lake/", package = "AEME")
  # Copy files from package into tempdir
  file.copy(aeme_dir, tmpdir, recursive = TRUE)
  path <- file.path(tmpdir, "lake")
  aeme_data <- AEME::yaml_to_aeme(path = path, "aeme.yaml")
  mod_ctrls <- read.csv(file.path(path, "model_controls.csv"))
  inf_factor = c("dy_cd" = 1, "glm_aed" = 1, "gotm_wet" = 1)
  outf_factor = c("dy_cd" = 1, "glm_aed" = 1, "gotm_wet" = 1)
  model <- c("glm_aed")
  aeme_data <- AEME::build_ensemble(path = path, aeme_data = aeme_data,
                                    model = model, mod_ctrls = mod_ctrls,
                                    inf_factor = inf_factor, ext_elev = 5,
                                    use_bgc = FALSE)

  # Load parameters
  utils::data("aeme_parameters", package = "aemetools")
  param <- aeme_parameters |>
    dplyr::filter(file != "wdr")

  # Function to calculate fitness
  fit <- function(df) {
    mean(df$model)
  }

  # Assign function to variable
  FUN_list <- list(HYD_temp = fit)

  # Set up control parameters for surface and bottom temperature
  ctrl <- create_control(method = "sa", N = 2^3, ncore = 2L,
                         parallel = TRUE,
                         vars_sim = list(
                                   surf_temp = list(var = "HYD_temp",
                                                    month = c(10:12, 1:3),
                                                    depth_range = c(0, 2)
                                                    ),
                                   bot_temp = list(var = "HYD_temp",
                                                   month = c(10:12, 1:3),
                                                   depth_range = c(10, 13)
                                                   )
                                   )
  )

  # Run sensitivity analysis AEME model
  ctrl <- sa_aeme(aeme_data = aeme_data, path = path, param = param,
                  model = model, ctrl = ctrl, mod_ctrls = mod_ctrls,
                  FUN_list = FUN_list)
}

}
