% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_param.R
\name{update_param}
\alias{update_param}
\title{Update parameter values in param based on best_pars}
\usage{
update_param(
  calib,
  param,
  na_value,
  aeme,
  replace = FALSE,
  quantile = 0.1,
  fit_col = "fit",
  best_pars
)
}
\arguments{
\item{calib}{dataframe; output from \code{\link{read_calib}}}

\item{param}{A data frame with parameters to update. Defaults to NULL. When
NULL, the parameter values are extracted from \code{calib$parameter_metadata}.}

\item{na_value}{A numeric value which corresponds to the NA value used in
the calibration.}

\item{aeme}{aeme; object. Defaults to NULL. When NULL, a dataframe of the
updated parameter values is returned. When provided, the updated parameter
values are added to the aeme object and the aeme object is returned.}

\item{replace}{Logical. If TRUE, the parameter values in the aeme object are
replaced with the updated values. Defaults to FALSE. Only used when aeme is
provided.}

\item{quantile}{The quantile to use for the top quantile of the fit_value.
Defaults to 0.1.}

\item{fit_col}{character; name of column containing fit values. Default is
\code{"fit"}.}

\item{best_pars}{A data frame with the best parameters from \code{\link{`get_param`}}.
Defaults to NULL. When NULL, \code{\link{`get_param`}} is called to get the
best parameters.}
}
\value{
data frame with updated parameter values for running the model with
\code{run_aeme_param}
}
\description{
Update parameter values in param based on best_pars
}
