% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_param.R
\name{update_param}
\alias{update_param}
\title{Update parameter values in param based on best_pars}
\usage{
update_param(
  calib,
  param,
  aeme,
  replace = FALSE,
  fit_col = "fit",
  best_pars,
  quantile = 0.1,
  na_value = NULL
)
}
\arguments{
\item{calib}{A list with the calibration results loaded using
\code{\link{read_calib}}.}

\item{param}{A data frame with parameters to update. Defaults to NULL. When
NULL, the parameter values are extracted from \code{calib$parameter_metadata}.}

\item{aeme}{aeme; object. Defaults to NULL. When NULL, a dataframe of the
updated parameter values is returned. When provided, the updated parameter
values are added to the aeme object and the aeme object is returned.}

\item{replace}{Logical. If TRUE, the parameter values in the aeme object are
replaced with the updated values. Defaults to FALSE. Only used when aeme is
provided.}

\item{fit_col}{character; name of column containing fit values. Default is
\code{"fit"}.}

\item{best_pars}{A data frame with the best parameters from \code{\link{get_param}}.
Defaults to NULL. When NULL, \code{\link{get_param}} is called to get the
best parameters.}

\item{quantile}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} The quantile to use for
the top quantile of the fit_value. Defaults to 0.1. This is no longer needed
and will be removed in a future version.}

\item{na_value}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Numeric. Penalty value
substituted for \code{NA} fit values, this is no longer needed as NA values
are now written to simulation_data in output of calib_aeme() and sa_aeme().
The argument will be removed in a future version.}
}
\value{
data frame with updated parameter values for running the model with
\code{\link{run_aeme_param}}
}
\description{
Update parameter values in param based on best_pars
}
