% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_param.R
\name{update_param}
\alias{update_param}
\title{Update parameter values in param based on best_pars}
\usage{
update_param(param, best_pars)
}
\arguments{
\item{param}{A data frame with parameters to update}

\item{best_pars}{A data frame with the best parameters from \code{get_param}}
}
\value{
data frame with updated parameter values for running the model with
\code{run_aeme_param}
}
\description{
Update parameter values in param based on best_pars
}
