% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_param.R
\name{update_param}
\alias{update_param}
\title{Update parameter values in param based on best_pars}
\usage{
update_param(
  param,
  calib,
  na_value,
  prob = 0.1,
  fit_col = "fit",
  best_pars = NULL
)
}
\arguments{
\item{param}{A data frame with parameters to update}

\item{calib}{A list with the calibration results loaded using
\code{\link{read_calib}}.}

\item{na_value}{A numeric value which corresponds to the NA value used in
the calibration.}

\item{fit_col}{character; name of column containing fit values. Default is
\code{"fit"}.}

\item{best_pars}{A data frame with the best parameters from \code{get_param}.
Defaults to NULL. When NULL, \code{get_param} is called to get the best
parameters.}
}
\value{
data frame with updated parameter values for running the model with
\code{run_aeme_param}
}
\description{
Update parameter values in param based on best_pars
}
