% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_csv_to_db.R
\name{write_csv_to_db}
\alias{write_csv_to_db}
\title{Write CSV to database}
\usage{
write_csv_to_db(path, file_name = "results.db")
}
\arguments{
\item{path}{filepath; where input files are located relative to `config`.}

\item{file_name}{The name of the database file to write to. Default is
"results.db".}
}
\value{
file path to the database.
}
\description{
Write CSV to database
}
